Add beginnings of Maestro event system / editor logic
Cocos doesn't have support for scriptable objects, so I'm modifying
the Maestro logic to route all key events through an AppSettings node
that is created during the Initializer scene.

I created several obsolete scripts that I thought would be used,
but which can't be used inside of Cocos, so like ModifiableEditorVariable
and others; I will remove those.

Next steps are to continue with implementation of ComplexPayload and
complex event triggers, in anticipation of being able to trigger
a sound effect and scene transition from a single button click,
agnostically and handled all thruogh the editor UI.