Begin adding support for dynamic scene data controlled via editor
We can now manipulate bools and ints using actionTriggerBehaviour in
conjunction with SceneData.

The SceneData script will keep track of variables using a variableKey
in the form of .txt asset.

If the variableKey is not present in SceneData, it will look for the
value in AppSettings.

This is the beginning of allowing us to have custom logic defined in the editor.

Next step is to add ConditionResponse action data, then do a cleanup of everything.