Fix bugs related to touch fork configuration
In order to prevent branch configurations from conflicting with each other
and the user getting stuck at the fork, made a few changes.

Namely, whenever a touch fork is activated, it will reset the destination on the
opposing branch, if it exists, so that if going from an inverted branch to a positive
branch, for example, the joiner won't immediately activate and ping pong the user
between the two branches, making them stuck.

This requires storing a copy of the touchForkExtents into an array on the Axis Monitor for
easy retrieval and comparison whenever a touchFork is activated (still need to clean up the
implementation for how the array is populated).

In addition, added the SEQUENCE_BOUNDARY_REACHED event back to the Joiner, and also made it
so the Axis Monitor will refresh the axes whenever a boundary is reached. The reason for this is
that sometimes, because in the step above we reset the destination on opposing branches,
sometimes a user could end up at the crossroads without a destination on the touchFork when there
actually should be one there. By firing SEQUENCE_BOUNDARY_REACHED, that allows us to activate
any touchForks if needed to update the destination branch and move the user through the
fork as expected.

Also, fixed a bug wherein the fork thresholds were being created incorrectly due to a typo
in AppSettings, and also added support for the forceForward and forceBackward states in the
Momentum Applier.